package structure re-org
