+++ projects/gateway2/translator/httproute/filterplugins/api/http_filter_plugin_api.go
@@ -3,7 +3,7 @@ package api
 import (
 	"context"
 
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions"
+	"github.com/solo-io/gloo/projects/gateway2/translator/plugins"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
@@ -12,7 +12,7 @@ type HTTPFilterPlugin interface {
 	// outputRoute.Options is guaranteed to be non-nil (TODO: confirm this still holds)
 	ApplyFilter(
 		ctx context.Context,
-		routeCtx *extensions.RouteContext,
+		routeCtx *plugins.RouteContext,
 		filter gwv1.HTTPRouteFilter,
 		outputRoute *v1.Route,
 	) error
