+++ projects/gateway2/translator/httproute/filterplugins/redirect/redirect_plugin.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	errors "github.com/rotisserie/eris"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions"
+	"github.com/solo-io/gloo/projects/gateway2/translator/plugins"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
@@ -18,7 +18,7 @@ func NewPlugin() *Plugin {
 
 func (p *Plugin) ApplyFilter(
 	ctx context.Context,
-	routeCtx *extensions.RouteContext,
+	routeCtx *plugins.RouteContext,
 	filter gwv1.HTTPRouteFilter,
 	outputRoute *v1.Route,
 ) error {
