+++ projects/gateway2/translator/httproute/filterplugins/redirect/redirect_plugin_test.go
@@ -2,9 +2,9 @@ package redirect_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/api"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/filtertests"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/redirect"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/api"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/filtertests"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/redirect"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
