+++ projects/gateway2/translator/httproute/filterplugins/urlrewrite/url_rewrite_test.go
@@ -2,9 +2,9 @@ package urlrewrite_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/api"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/filtertests"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/filterplugins/urlrewrite"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/api"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/filtertests"
+	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/urlrewrite"
 	v3 "github.com/solo-io/gloo/projects/gloo/pkg/api/external/envoy/type/matcher/v3"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"google.golang.org/protobuf/types/known/wrapperspb"
