+++ projects/gateway2/translator/plugins/plugins.go
@@ -1,4 +1,4 @@
-package extensions
+package plugins
 
 import (
 	"context"
