+++ projects/gateway2/xds/xds_syncer.go
@@ -11,7 +11,7 @@ import (
 	"github.com/solo-io/gloo/projects/gateway2/query"
 	"github.com/solo-io/gloo/projects/gateway2/reports"
 	gloot "github.com/solo-io/gloo/projects/gateway2/translator"
-	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/routeregistry"
+	"github.com/solo-io/gloo/projects/gateway2/translator/plugins/registry"
 	gloo_solo_io "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	v1snap "github.com/solo-io/gloo/projects/gloo/pkg/api/v1/gloosnapshot"
 	"github.com/solo-io/gloo/projects/gloo/pkg/plugins"
@@ -157,7 +157,7 @@ func (s *XdsSyncer) Start(
 			return
 		}
 		queries := query.NewData(s.cli, s.scheme)
-		routePlugins := routeregistry.NewRoutePluginRegistry(queries)
+		routePlugins := registry.NewRoutePluginRegistry(queries)
 		t := gloot.NewTranslator(*routePlugins)
 		proxies := gloo_solo_io.ProxyList{}
 		rm := reports.NewReportMap()
